Update index.html to improve a few things
change the name of the Ensemble tab to Results

Add a toggle on the ensemble/results tab that allows the user to change the data being shown and routed through the trees from train to test

Only show Train accuracy when train is toggled (Should be default) and then show test and the difference from train in each tree tab once test has been toggled in the results tab. Encourage the user to change to test once they are happy with the trees they have built.

Include the names of passengers in the data but not as an optional variable in the trees

Include the title in the data tab and the title as a option with a selectable drop down for title in the tree tabs

The spacing of the terminal nodes is too close together and the lines branching from the nodes above should connect to the nodes below. rework this look especially for the final nodes

The information in the node denoting which passengers are currently in each node should be above the user selection

Ensure the < and > logic is correct as in some nodes i believe the greater than and less than operators were working as an equals.